Improving README and making the dataset loading function public
