+++ sdgym/__init__.py
@@ -11,7 +11,9 @@ __license__ = 'MIT'
 __version__ = '0.1.0-dev'
 
 from sdgym.benchmark import benchmark
+from sdgym.data import load_dataset
 
 __all__ = [
-    'benchmark'
+    'benchmark',
+    'load_dataset'
 ]
